PIA/requirements.txt
arrow==1.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
docopt==0.6.2
future==0.18.2
hunter==3.3.8
idna==3.3
mime==0.1.0
MouseInfo==0.1.3
nmap==0.0.1
Pillow==8.4.0
pipreqs==0.4.11
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
python-nmap==0.7.1
pytweening==1.0.4
PyYAML==6.0
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
times==0.7
turtle==0.0.1
urllib3==1.26.7
webscraping==1.7.1
XlsxWriter==3.0.2
yarg==0.1.9
